% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampseq.qc-package.R
\docType{package}
\name{ampseq.qc-package}
\alias{ampseq.qc}
\alias{ampseq.qc-package}
\title{ampseq.qc: QC Summary Shiny App}
\description{
A Shiny app for visualizing QC summary data from ampseq data.
}
\author{
\strong{Maintainer}: Kathryn Murie \email{kathryn.murie@ucsf.edu}

Authors:
\itemize{
  \item Maxwell Murphy \email{maxwell.murphy@ucsf.edu} [contributor]
}

}
\keyword{internal}
